Enable caching of PermutationIterator methods
